+++ src/iqm/cirq_iqm/__init__.py
@@ -25,3 +25,6 @@ except PackageNotFoundError:
     __version__ = 'unknown'
 finally:
     del version, PackageNotFoundError
+
+from .iqm_gates import *
+from .transpiler import transpile_insert_moves_into_circuit