+++ src/iqm/cirq_iqm/iqm_sampler.py
@@ -28,21 +28,8 @@ import cirq
 import numpy as np
 
 from iqm.cirq_iqm.devices.iqm_device import IQMDevice, IQMDeviceMetadata
-from iqm.cirq_iqm.iqm_operation_mapping import map_operation
-from iqm.iqm_client import Circuit, HeraldingMode, IQMClient, JobAbortionError, RunRequest
-
-
-def serialize_circuit(circuit: cirq.Circuit) -> Circuit:
-    """Serializes a quantum circuit into the IQM data transfer format.
-
-    Args:
-        circuit: quantum circuit to serialize
-
-    Returns:
-        data transfer object representing the circuit
-    """
-    instructions = tuple(map(map_operation, circuit.all_operations()))
-    return Circuit(name='Serialized from Cirq', instructions=instructions)
+from iqm.cirq_iqm.serialize import serialize_circuit
+from iqm.iqm_client import HeraldingMode, IQMClient, JobAbortionError, RunRequest
 
 
 class IQMSampler(cirq.work.Sampler):